refactor(integration): move kotlinx & spring extensions to extension package
Relocate the kotlinx-serialization and spring IrExtension classes from the
`emit` package to a dedicated `extension` package, updating imports in the
affected tests.
