+++ src/integration/spring/src/jvmTest/kotlin/community/flock/wirespec/integration/spring/java/emit/SpringJavaIrExtensionTest.kt
@@ -13,8 +13,8 @@ import community.flock.wirespec.compiler.core.parse.ast.AST
 import community.flock.wirespec.compiler.utils.NoLogger
 import community.flock.wirespec.compiler.utils.noLogger
 import community.flock.wirespec.emitters.java.JavaIrEmitter
-import community.flock.wirespec.integration.spring.emit.SpringMappingAnnotationsSupportExtension
-import community.flock.wirespec.integration.spring.emit.SpringMappingNativeSupportExtension
+import community.flock.wirespec.integration.spring.extension.SpringMappingAnnotationsSupportExtension
+import community.flock.wirespec.integration.spring.extension.SpringMappingNativeSupportExtension
 import community.flock.wirespec.ir.extension.applyExtensions
 import kotlinx.io.buffered
 import kotlinx.io.files.Path
